+++ src/main.cpp/DisplayInit.h
@@ -21,10 +21,10 @@
 // Button and potentiometer configuration
 
 
-extern uint16_t mappedValue;
-extern uint16_t lastMappedValue;
-extern uint8_t currentpage;
-extern uint8_t screencount;
+extern int mappedValue;
+extern int lastMappedValue;
+extern int currentpage;
+extern int screencount;
 
 
 // Array of function pointers for page handling
